chore: updated README.md and small refactor
